Renamed test/service.py to test/_service.py
